# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TerraCare API",
    "version": "1.0.0",
    "description": "Terariumų, augintinių ir jų priežiūros valdymo API."
  },
  "servers": [
    {
      "url": "/",
      "description": "Dabartinis serveris"
    }
  ],
  "paths": {
    "/pets/{id}/profile": {
      "get": {
        "tags": [
          "Augintiniai"
        ],
        "summary": "Gauti augintinio profilį",
        "description": "Viename atsakyme sujungiama augintinio informacija, jo terariumas ir priežiūros istorija. Jei priežiūros įrašų nėra, care_records yra tuščias masyvas.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Augintinio ID.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Augintinio profilis iš trijų susijusių esybių",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "pet",
                    "terrarium",
                    "care_records"
                  ],
                  "properties": {
                    "pet": {
                      "$ref": "#/components/schemas/Pet"
                    },
                    "terrarium": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "temperature_min": {
                          "type": "number"
                        },
                        "temperature_max": {
                          "type": "number"
                        },
                        "humidity_min": {
                          "type": "number"
                        },
                        "humidity_max": {
                          "type": "number"
                        },
                        "last_cleaned": {
                          "type": "string",
                          "format": "date",
                          "nullable": true
                        },
                        "notes": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "care_records": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CareRecord"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas augintinio ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Netinkamas augintinio ID."
                }
              }
            }
          },
          "404": {
            "description": "Augintinis nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Augintinis nerastas."
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti augintinio profilio."
                }
              }
            }
          }
        }
      }
    },
    "/terrariums/{terrariumId}/pets/{petId}/care-records": {
      "get": {
        "tags": [
          "Priežiūros įrašai"
        ],
        "summary": "Gauti terariumo augintinio priežiūros įrašus",
        "description": "Tikrinama, ar terariumas egzistuoja ir ar augintinis priklauso būtent jam. Grąžinama augintinio priežiūros istorija nuo naujausių įrašų. Jei istorija tuščia, grąžinamas tuščias masyvas.",
        "parameters": [
          {
            "name": "terrariumId",
            "in": "path",
            "required": true,
            "description": "Terariumo ID.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "petId",
            "in": "path",
            "required": true,
            "description": "Šiam terariumui priklausančio augintinio ID.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Augintinio priežiūros įrašų sąrašas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CareRecord"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas terariumo arba augintinio ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Terariumo ir augintinio ID turi būti teigiami sveikieji skaičiai."
                }
              }
            }
          },
          "404": {
            "description": "Terariumas nerastas arba augintinis šiame terariume nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "terrariumNotFound": {
                    "summary": "Terariumas neegzistuoja",
                    "value": {
                      "error": "Terariumas nerastas."
                    }
                  },
                  "petNotInTerrarium": {
                    "summary": "Augintinis neegzistuoja arba priklauso kitam terariumui",
                    "value": {
                      "error": "Augintinis šiame terariume nerastas."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti augintinio priežiūros įrašų."
                }
              }
            }
          }
        }
      }
    },
    "/terrariums/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Terariumo ID",
          "schema": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          }
        }
      ],
      "get": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Gauti terariumą pagal ID",
        "responses": {
          "200": {
            "description": "Terariumo informacija",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terrarium"
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas terariumo ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Terariumas nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Terariumas nerastas."
                }
              }
            }
          },
          "500": {
            "$ref": "#/paths/~1terrariums/get/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Atnaujinti terariumą",
        "description": "Pakeičia terariumo informaciją. Reikia pateikti pavadinimą ir visas temperatūros bei drėgmės ribas. Nepateikus last_cleaned ar notes, jie nustatomi į null.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerrariumInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terariumas atnaujintas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terrarium"
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas ID arba netaisyklingas JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/404"
          },
          "422": {
            "$ref": "#/paths/~1terrariums/post/responses/422"
          },
          "500": {
            "$ref": "#/paths/~1terrariums/get/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Ištrinti terariumą",
        "description": "Galima ištrinti tik terariumą, kuriame nėra augintinių.",
        "responses": {
          "204": {
            "description": "Terariumas ištrintas. Atsakymo turinio nėra."
          },
          "400": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/400"
          },
          "404": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/404"
          },
          "409": {
            "description": "Terariume yra augintinių",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Terariumo ištrinti negalima, nes jame yra augintinių."
                }
              }
            }
          },
          "500": {
            "$ref": "#/paths/~1terrariums/get/responses/500"
          }
        }
      }
    },
    "/terrariums/{id}/pets": {
      "get": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Gauti konkretaus terariumo augintinius",
        "description": "Hierarchinis metodas. Jei terariumas egzistuoja, bet neturi augintinių, grąžinamas tuščias masyvas. Jei terariumo nėra, grąžinamas 404.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Terariumo ID",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pasirinkto terariumo augintinių sąrašas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/400"
          },
          "404": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/404"
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti terariumo augintinių."
                }
              }
            }
          }
        }
      }
    },
...
    "/pets/{id}": {
      "put": {
        "tags": [
          "Augintiniai"
        ],
        "summary": "Atnaujinti augintinį",
        "description": "Pakeičia augintinio kortelę. Nepateikus arrival_date arba notes, šie laukai nustatomi į null.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "$ref": "#/paths/~1pets/post/requestBody"
        },
        "responses": {
          "200": {
            "description": "Augintinis atnaujintas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas ID arba netaisyklingas JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Netinkamas augintinio ID."
                }
              }
            }
          },
          "404": {
            "description": "Augintinis nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Augintinis nerastas."
                }
              }
            }
          },
          "422": {
            "description": "Netinkami duomenys arba nurodytas terariumas neegzistuoja",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nurodytas terariumas neegzistuoja."
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko atnaujinti augintinio."
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Augintiniai"
        ],
        "summary": "Ištrinti augintinį",
        "description": "Augintinį galima ištrinti tik tada, kai jis neturi priežiūros įrašų.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Augintinis ištrintas. Atsakymo turinio nėra."
          },
          "400": {
            "description": "Netinkamas augintinio ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Netinkamas augintinio ID."
                }
              }
            }
          },
          "404": {
            "description": "Augintinis nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Augintinis nerastas."
                }
              }
            }
          },
          "409": {
            "description": "Augintinis turi priežiūros įrašų",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Augintinio ištrinti negalima, nes jis turi priežiūros įrašų."
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko ištrinti augintinio."
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Augintiniai"
        ],
        "summary": "Gauti augintinį pagal ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Augintinio ID",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Augintinio informacija",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas augintinio ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Augintinio ID turi būti teigiamas sveikasis skaičius."
                }
              }
            }
          },
          "404": {
            "description": "Augintinis nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Augintinis nerastas."
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti augintinio."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Terrarium": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Varlių terariumas"
          },
          "temperature_min": {
            "type": "string",
            "example": "22.00"
          },
          "temperature_max": {
            "type": "string",
            "example": "26.00"
          },
          "humidity_min": {
            "type": "string",
            "example": "60.00"
          },
          "humidity_max": {
            "type": "string",
            "example": "80.00"
          },
          "last_cleaned": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Demonstracinis terariumas."
          }
        }
      },
...
      "CareRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "pet_id": {
            "type": "integer",
            "example": 1
          },
          "care_type": {
            "type": "string",
            "enum": [
              "feeding",
              "observation",
              "other"
            ]
          },
          "performed_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "description": "Laiką automatiškai nustato serveris."
          },
          "notes": {
            "type": "string",
            "example": "Augintinis aktyvus, pakitimų nepastebėta."
          }
        }
      },
...
      "Pet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "terrarium_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Pupa"
          },
          "species": {
            "type": "string",
            "example": "Litoria caerulea"
          },
          "arrival_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-09-01"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Aktyvi vakare."
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Nepavyko gauti augintinių sąrašo."
          }
        }
      }
    }
  }
}